finir de labo5-ex3

+++ inf111-labo5/src/com/atoudeft/tdas/Point.java
@@ -30,7 +30,7 @@ public class Point {
 
     @Override
     public String toString() {
-        return "("+ abs + "," + ord + ")";
+        return "("+ (int)abs + "," + (int)ord + ")";
     }
 
     public double dist(Point point) {
